refactor: move layout files from core to lib
