--- core/ui/layout/ScrollableFlexboxFiller.tsx
@@ -1,11 +1,0 @@
-import { ChildrenProp, UiProps } from '@lib/ui/props'
-
-export const ScrollableFlexboxFiller = ({
-  children,
-  className,
-  style,
-}: ChildrenProp & UiProps) => (
-  <div className={`flex-1 relative ${className}`} style={style}>
-    <div className="absolute inset-0 overflow-auto">{children}</div>
-  </div>
-)
