+++ core/ui/vaultsOrganisation/components/VaultListItem.tsx
@@ -2,7 +2,7 @@ import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
 import { useSetCurrentVaultIdMutation } from '@core/ui/storage/currentVaultId'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
 import { getVaultId } from '@core/ui/vault/Vault'
-import { ListItem } from '@lib/list/item/ListItem'
+import { ListItem } from '@lib/ui/list/item/ListItem'
 
 import { VaultDescription } from './VaultDescription'
