+++ core/ui/vaultsOrganisation/folders/components/FolderListItem.tsx
@@ -1,6 +1,6 @@
-import { ListItem } from '@lib/list/item/ListItem'
 import { FolderIcon } from '@lib/ui/icons/FolderIcon'
 import { HStack } from '@lib/ui/layout/Stack'
+import { ListItem } from '@lib/ui/list/item/ListItem'
 
 import { useCoreNavigate } from '../../../navigation/hooks/useCoreNavigate'
