+++ core/ui/vaultsOrganisation/folders/manage/ManageVaultFolders.tsx
@@ -3,11 +3,11 @@ import {
   useVaultFolders,
 } from '@core/ui/storage/vaultFolders'
 import { VaultFolder } from '@core/ui/vault/VaultFolder'
+import { DnDList } from '@lib/ui/dnd/DnDList'
 import {
   DnDItemContainer,
   DnDItemHighlight,
-} from '@lib/list/item/DnDItemContainer'
-import { DnDList } from '@lib/ui/dnd/DnDList'
+} from '@lib/ui/list/item/DnDItemContainer'
 import { isEmpty } from '@lib/utils/array/isEmpty'
 import { sortEntitiesWithOrder } from '@lib/utils/entities/EntityWithOrder'
 import { getNewOrder } from '@lib/utils/order/getNewOrder'
