--- lib/list/package.json
@@ -1,11 +1,0 @@
-{
-    "name": "@lib/list",
-    "version": "0.0.1",
-    "private": true,
-    "scripts": {
-        "typecheck": "tsc --noEmit"
-    },
-    "dependencies": {
-        "styled-components": "^6.1.16"
-    }
-}
